requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
joblib
scikit-learn
protobuf<4.21.0
